move NoAnonymization class near the interface in AnonymizationProvider module
