+++ bin/mysql-dump-anonymize.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use PayU\MysqlDumpAnonymizer\Application\Application;
+namespace PayU\MysqlDumpAnonymizer\Application;
 
 require_once dirname(__DIR__).'/vendor/autoload.php';
