+++ lib/AnonymizationProvider/AnonymizationProvider.php
@@ -6,8 +6,6 @@ namespace PayU\MysqlDumpAnonymizer\AnonymizationProvider;
 
 final class AnonymizationProvider implements AnonymizationProviderInterface
 {
-    public const NO_ANONYMIZATION = 'NoAnonymization';
-
     /** @var array  */
     private array $tablesAction;
 
@@ -61,10 +59,6 @@ final class AnonymizationProvider implements AnonymizationProviderInterface
 
     public function isAnonymization(ValueAnonymizerInterface $valueAnonymizer): bool
     {
-        $className = get_class($valueAnonymizer);
-        if ($shortClassName = strrchr($className, "\\")) {
-            $className = substr($shortClassName, 1);
-        }
-        return $className !== self::NO_ANONYMIZATION;
+        return !$valueAnonymizer instanceof NoAnonymization;
     }
 }
